Maç Skoru kaydetme

+++ backend/service/internal/delivery/http/handler/matchhandler/scoreValidator.go
@@ -57,7 +57,7 @@ func RegisterSetValidations() {
 	}
 }
 
-func abs(x int64) int64 {
+func abs(x int8) int8 {
 	if x < 0 {
 		return -x
 	}
